% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectations.R
\name{assign_depart_hard}
\alias{assign_depart_hard}
\title{Assign departments to students according to their wishes, regardless of the number of places}
\usage{
assign_depart_hard(df, selection, wish_number)
}
\arguments{
\item{df}{The data frame with the students and their wishes}

\item{selection}{The indices of students to assign}

\item{wish_number}{The number of the wish to assign}
}
\value{
A list with (list: The input data frame with affected departments, fails: The indices of students that could not be assigned)
}
\description{
Assign departments to students according to their wishes, regardless of the number of places
}
