% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectations.R
\name{assign_depart_soft}
\alias{assign_depart_soft}
\title{Assign departments to students according to their wishes, looking for the first wish that is not full}
\usage{
assign_depart_soft(df, selection, capacities)
}
\arguments{
\item{df}{The data frame with the students and their wishes}

\item{selection}{The indices of students to assign}

\item{capacities}{The remaining capacities of the departments}
}
\value{
A list with (list: The input data frame with affected departments, fails: The indices of students that could not be assigned)
}
\description{
Assign departments to students according to their wishes, looking for the first wish that is not full
}
