% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectations.R
\name{assign_erase}
\alias{assign_erase}
\title{Erase all affected departments}
\usage{
assign_erase(df, selection, sessions = FALSE)
}
\arguments{
\item{df}{The data frame with the students and their wishes}

\item{selection}{The indices of students to erase}

\item{sessions}{Whether to erase affected sessions or departments}
}
\value{
A list with (list: The input data frame with erased departments, fails: An empty vector because this can never fail)
}
\description{
Erase all affected departments
}
