% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectations.R
\name{assign_session_auto}
\alias{assign_session_auto}
\title{Automatically assign sessions to students based on their department affectations}
\usage{
assign_session_auto(df, selection, capacities)
}
\arguments{
\item{df}{The data frame with the students and affected departements}

\item{selection}{The indices of students to assign}

\item{capacities}{The total capacities of the departments over 3 sessions}
}
\value{
A list with (list: The input data frame with affected sessions, fails: The indices of students that could not be assigned)
}
\description{
Automatically assign sessions to students based on their department affectations
}
