% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectations.R
\name{assign_session_manual}
\alias{assign_session_manual}
\title{Manually assign a department to students' sessions}
\usage{
assign_session_manual(df, selection, depart, session)
}
\arguments{
\item{df}{The data frame with the students and affected departements}

\item{selection}{The indices of students to assign}

\item{depart}{The department to assign}

\item{session}{The session to assign it to}
}
\value{
A list with (list: The input data frame with affected sessions, fails: The indices of students that could not be assigned)
}
\description{
Manually assign a department to students' sessions
}
