% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacities.R
\name{calculate_new_capacities}
\alias{calculate_new_capacities}
\title{Calculate the new capacities avec each changes of the df}
\usage{
calculate_new_capacities(df, capacities)
}
\arguments{
\item{df}{The reactive data frame of students's wishes and affectations}

\item{capacities}{The reactive table of each departments capacities}
}
\description{
Calculate the new capacities avec each changes of the df
}
