% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacities.R
\name{calculate_session_capacities}
\alias{calculate_session_capacities}
\title{Calculate the new department capacities for each session}
\usage{
calculate_session_capacities(df, capacities)
}
\arguments{
\item{df}{The reactive data frame of students's wishes and affectations}

\item{capacities}{The reactive table of each departments capacities}
}
\description{
Calculate the new department capacities for each session
}
