% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-data-generation.R
\docType{data}
\name{classic_answer}
\alias{classic_answer}
\title{Possible answers}
\format{
An object of class \code{character} of length 7.
}
\usage{
classic_answer
}
\description{
Possible answers
}
\keyword{datasets}
