% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{color_departments_column}
\alias{color_departments_column}
\title{Create a JavaScript function to color the departments column according to the worst wish of a student}
\usage{
color_departments_column(df, filtered)
}
\arguments{
\item{df}{The full main data frame}

\item{filtered}{The filtered data frame that will be displayed}
}
\value{
A JavaScript function as a vector of lines
}
\description{
Create a JavaScript function to color the departments column according to the worst wish of a student
}
