% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{color_session_cells}
\alias{color_session_cells}
\title{Color the cells of the session columns based on the wishes}
\usage{
color_session_cells(wb, df)
}
\arguments{
\item{wb}{The workbook that will be exported}

\item{df}{The main data frame}
}
\description{
Color the cells of the session columns based on the wishes
}
