% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-data-generation.R
\docType{data}
\name{df}
\alias{df}
\title{generates a dataframe of the students with random informations (names/lastnames/sector)}
\format{
An object of class \code{data.frame} with 200 rows and 3 columns.
}
\usage{
df
}
\value{
dataframe with student's infos
}
\description{
generates a dataframe of the students with random informations (names/lastnames/sector)
}
\keyword{datasets}
