% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{display_tables}
\alias{display_tables}
\title{Function used by server to display tables}
\usage{
display_tables(input, output, df)
}
\arguments{
\item{input}{Input data from the frontend}

\item{output}{Output data the frontend will receive}

\item{df}{The reactive data frame of students's wishes and affectations}
}
\description{
Function used by server to display tables
}
