% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{filter_for_table}
\alias{filter_for_table}
\title{Filter the data frame for the specified table}
\usage{
filter_for_table(df, table, input)
}
\arguments{
\item{df}{The data frame to filter}

\item{table}{The name of the table to filter for}

\item{input}{The input data from the frontend}
}
\description{
Filter the data frame for the specified table
}
