% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-data-generation.R
\name{final_dataframe}
\alias{final_dataframe}
\title{applies wishes_generation to each line of the dataframe of the students, and adds the generated wishes in the dataframe
fusion the dataframes to create the final one}
\usage{
final_dataframe(df)
}
\arguments{
\item{df}{dataframe with the student's infos}
}
\value{
a dataframe with the wishes
}
\description{
applies wishes_generation to each line of the dataframe of the students, and adds the generated wishes in the dataframe
fusion the dataframes to create the final one
}
