% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_selection}
\alias{get_selection}
\title{Get the selection from the specified table}
\usage{
get_selection(df, table, input)
}
\arguments{
\item{df}{The main data frame}

\item{table}{The name of the table to select from}

\item{input}{The input data from the frontend}
}
\description{
Get the selection from the specified table
}
