% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectations.R
\name{handle_affectations}
\alias{handle_affectations}
\title{Function used by server to handle affectations}
\usage{
handle_affectations(
  input,
  output,
  df,
  capacities,
  remaining_depart_capacities,
  remaining_session_capacities
)
}
\arguments{
\item{input}{Input data from the frontend}

\item{output}{Output data the frontend will receive}

\item{df}{The reactive data frame of students's wishes and affectations}

\item{capacities}{The total capacities of the departments over 3 sessions}

\item{remaining_depart_capacities}{The remaining capacities of the departments}

\item{remaining_session_capacities}{The remaining capacities of the sessions}
}
\description{
Function used by server to handle affectations
}
