% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{operation_tooltip}
\alias{operation_tooltip}
\title{Custom tooltip with preset options for the operation buttons}
\usage{
operation_tooltip(
  target,
  text,
  placement = "bottom",
  options = list(delay = list(show = 250, hide = 0)),
  ...
)
}
\arguments{
\item{target}{The target element for the tooltip}

\item{text}{The text of the tooltip}

\item{placement}{The placement of the tooltip}

\item{options}{Additional options for the tooltip}

\item{...}{Additional arguments for the tooltip}
}
\value{
A bslib tooltip object
}
\description{
Custom tooltip with preset options for the operation buttons
}
