% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{process_data}
\alias{process_data}
\title{Processes the file data to match our data structure}
\usage{
process_data(file_data, data_source)
}
\arguments{
\item{file_data}{Raw data from the file}

\item{data_source}{The string of the data source (Moodle or ADSPlanner)}
}
\value{
The final data frame
}
\description{
Processes the file data to match our data structure
}
