% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-data-generation.R
\name{random_data}
\alias{random_data}
\title{Applies wishes_generation to each line of the dataframe of the students, and adds the generated wishes in the dataframe}
\usage{
random_data()
}
\value{
a dataframe with the wishes
}
\description{
Applies wishes_generation to each line of the dataframe of the students, and adds the generated wishes in the dataframe
}
