% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Backend server logic}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Input data from the frontend}

\item{output}{Output data the frontend will receive}

\item{session}{The Shiny session object}
}
\description{
Backend server logic
}
