% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{synthesize_wishes}
\alias{synthesize_wishes}
\title{Synthesizes a student's wishes from multiple columns per filiere to 7 columns for all filieres}
\usage{
synthesize_wishes(wishes)
}
\arguments{
\item{wishes}{The wishes columns}
}
\value{
A vector with the filiere and the 7 wishes (NA if no wish)
}
\description{
Synthesizes a student's wishes from multiple columns per filiere to 7 columns for all filieres
}
