% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectations.R
\name{targeted_affectation}
\alias{targeted_affectation}
\title{assign certain students to a certain departement, deleting it from a certain departement}
\usage{
targeted_affectation(df, selection, old_depart, new_depart)
}
\arguments{
\item{df}{The data frame with the students and their wishes}

\item{selection}{The indices of students to assign}

\item{old_depart}{The departement the students were assigned}

\item{new_depart}{The departement to which the students will be assigned}
}
\value{
The input data frame with affected departments
}
\description{
assign certain students to a certain departement, deleting it from a certain departement
}
