% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-data-generation.R
\name{wishes_generation}
\alias{wishes_generation}
\title{Generates student's wishs in fonction of their sector (MICA/EMIR/CLASSIQUE)}
\usage{
wishes_generation(sector)
}
\arguments{
\item{sector}{"CLASSIQUE", "EMIR" ou "MICA".}
}
\value{
a vector with attributed wishes
}
\description{
Generates student's wishs in fonction of their sector (MICA/EMIR/CLASSIQUE)
}
