% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacities.R
\name{handle_capacities}
\alias{handle_capacities}
\title{Handle capacities on server}
\usage{
handle_capacities(input, output, df, capacities, remaining_capacities)
}
\arguments{
\item{input}{Input data from the frontend}

\item{output}{Output data the frontend will receive}

\item{df}{The reactive data frame of students's wishes and affectations}

\item{capacities}{The reactive table of each departments capacities}

\item{remaining_capacities}{The reactive table of the remaining capacities in each department}
}
\description{
Handle capacities on server
}
