% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{parse_file}
\alias{parse_file}
\title{Parses the file to build a clean data frame}
\usage{
parse_file(file_path)
}
\arguments{
\item{file_path}{The path to the file to parse}
}
\value{
A data frame with the cleaned data
}
\description{
Parses the file to build a clean data frame
}
