% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_file}
\alias{read_file}
\title{Extracts raw data from the provided file path (.xlsx or .ods)}
\usage{
read_file(file_path)
}
\arguments{
\item{file_path}{The path to the file to extract data from}
}
\value{
A list containing the file data and the sheet names
}
\description{
Extracts raw data from the provided file path (.xlsx or .ods)
}
