% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_file}
\alias{read_file}
\title{Reads the file and selects the needed columns}
\usage{
read_file(file_path)
}
\arguments{
\item{file_path}{The path to the file to parse}
}
\value{
The file as a data frame
}
\description{
Reads the file and selects the needed columns
}
