% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{write_output}
\alias{write_output}
\title{Output data to an Excel file}
\usage{
write_output(df, capacities, file)
}
\arguments{
\item{df}{The data frame to write}

\item{capacities}{The capacities of each department}

\item{file}{The path to the file to write}
}
\description{
Output data to an Excel file
}
